Move Thread to types
