+++ packages/fake-backends/src/Model.ts
@@ -1,12 +1,13 @@
 // Copyright (c) Microsoft Corporation.
 // Licensed under the MIT license.
 
-import { ChatParticipant, ChatMessage, ChatThreadProperties, ChatMessageReadReceipt } from '@azure/communication-chat';
+import { ChatParticipant, ChatMessage } from '@azure/communication-chat';
 import { CommunicationIdentifier } from '@azure/communication-common';
 import { toFlatCommunicationIdentifier } from '@internal/acs-ui-common';
 import { EventEmitter } from 'events';
 import produce from 'immer';
 import { ThreadEventEmitter } from './ThreadEventEmitter';
+import { Thread } from './types';
 
 export class Model {
   private threads: { [key: string]: Thread } = {};
@@ -46,13 +47,6 @@ export class Model {
       });
     }
   }
-}
-
-export interface Thread extends ChatThreadProperties {
-  version: number;
-  participants: ChatParticipant[];
-  messages: ChatMessage[];
-  readReceipts: ChatMessageReadReceipt[];
 }
 
 export const bumpMessageVersion = (message: ChatMessage): void => {
